requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
BingImageCreator==0.5.0
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
EdgeGPT==0.13.2
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
prompt-toolkit==3.0.39
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
regex==2023.8.8
requests==2.31.0
rich==13.5.2
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.4.1
twilio==8.7.0
urllib3==2.0.4
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
